feat: revise the header component
헤더 컴포넌트를 통합 및 수정한다
