+++ index.html
@@ -11,7 +11,7 @@
     <header class="header">
       <div class="header__title">
         <img src="./asset/symbol/logo.svg" alt="logo" />
-        <a onclick="window.location.reload()">뉴스스탠드</a>
+        <a>뉴스스탠드</a>
       </div>
       <div class="header__date">
         <span class="date"></span>
@@ -35,6 +35,7 @@
           </div>
         </div>
       </section>
+
       <section class="press">
         <div class="press__header">
           <div class="tap">
